Upgrade AOSP Dialer to Android Q, add necessary permissions in application side.
1.Protect Device Identifiers: Starting in Android Q, apps must have the READ_PRIVILEGED_PHONE_STATE privileged permission in order to access the device's non-resettable identifiers
IMEI and serial is used for the *#06# dialog. OEMs needs this to run factory tests. Perhaps it is something that should be moved to telephony, but as the current state dialer is showing it.

2.Limit pre-installed app network permission:Preinstalled apps that want to access restricted network needs to declare the CONNECTIVITY_USE_RESTRICTED_NETWORKS permission.
Restricted network is used for VVM access.

3.Notification fullscreen intent attribution, Developers need to add a uses-permission tag for android.permission.USE_FULL_SCREEN_INTENT to their manifest
Full screen intent is used for incoming call notification

Bug: 132081872 132080738 132114409
Test: Compile and load build.

Change-Id: Iec458cef1ff0c5a4180b438b8ba1a5159bdd8e92
(cherry picked from commit e9f21732d5253117e6546d792ef4b931775c6c78)
CRs-Fixed: 2576855