IMS: Do not show wrong prompt in case of VT upgrade request.
When an incoming call is received on the other sub, prompt
"Answering this call will end your other call" is shown.
Once the call on other sub is active and an incoming modify
request is received for the call, the same prompt is wrongly
shown again.

Fix this by checking for incoming upgrade request and returning
false from the api which decides whether prompt should be shown.

Change-Id: I258dafd0d4ddcf11e34fa627bf80a203484fee8e
CRs-Fixed: 2775109